Stopped adding documents to context. Tests failing.
